Introduce page to single node in android
This will somehow act as 'title' was used before, in the sense that it
will be used to map to a js class when refreshing the app or for example
pushing a new screen.
The actual visual title should be set elsewhere and is rendered as
'todo' for the time being.